requirements.txt
streamlit==1.44.1
requests==2.32.3
python-dotenv==1.1.0
structlog==25.2.0
pydantic-settings==2.8.1
langchain== 0.3.23
langchain-groq==0.3.2
langchain-openai==0.3.12
langchain-community==0.3.21
psycopg2-binary==2.9.10
